Se agregan cambios y mejoras

+++ TP1/funciones.h
@@ -1,47 +1,47 @@
 /** \brief La funcion realiza la suma de los numeros dados por parametros
  *
- * \param  Numero1 variable que guarda el primer numero
- * \param  Numero2 variable que guarda el primer numero
+ * \param  x variable que guarda el primer numero
+ * \param  y variable que guarda el primer numero
  *
  * \return Retorna la suma de los dos numeros
  */
-unsigned int suma (int x, int y);
+float suma (float x, float y);
 
 
 /** \brief La funcion realiza la resta de los numeros dados por parametros
  *
- * \param  Numero1 variable que guarda el primer numero
- * \param  Numero2 variable que guarda el primer numero
+ * \param  x variable que guarda el primer numero
+ * \param  y variable que guarda el primer numero
  *
  * \return Retorna la resta de los dos numeros
  */
-unsigned int resta (int x, int y);
+float resta (float x, float y);
 
 
 /** \brief La funcion realiza la divison de los numeros dados por parametros
  *
- * \param Numero1 variable que guarda el primer numero
- * \param Numero2 variable que guarda el primer numero
+ * \param x variable que guarda el primer numero
+ * \param y variable que guarda el primer numero
  *
- * \return Retorna la division del Numero1 sobre el Numero2
+ * \return Retorna la division del numero X sobre el numero Y
  */
-float dividido (int x, int y);
+float dividido (float x, float y);
 
 
 /** \brief La funcion realiza la multiplicacion de los numeros dados por parametros
  *
- * \param  Numero1 variable que guarda el primer numero
- * \param  Numero2 variable que guarda el primer numero
+ * \param  x variable que guarda el primer numero
+ * \param  y variable que guarda el primer numero
  *
  * \return Retorna la multiplicacion de los dos numeros dados
  */
-unsigned int multiplicar (int x, int y);
+float multiplicar (float x, float y);
 
 
 /** \brief La funcion realiza la factorizacion del numero dado, el mismo no puede ser 0 o negativo
  *
- * \param  Numero1 variable que guarda el primer numero
+ * \param  x variable que guarda el primer numero
  *
  * \return Retorna la factorizacion del numero dado
  */
-unsigned int factorizar (int x);
+float factorizar (float x);
